pyproject.toml
[project]
name = "ocrmypdf-appleocr"
dynamic = ["version"]
description = "Plugin to run OCRmyPDF with Apple Vision Framework OCR engine"
authors = [{ name = "Masahiro Kiyota", email = "hiro@juzbox.com" }]
dependencies = [
  "pyobjc-core",
  "pyobjc-framework-Vision",
  "pyobjc-framework-Cocoa",
  "langdetect",
  "ocrmypdf>=14.2.1",
  "Pillow>=10.0.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["pdf", "ocr", "optical character recognition", "apple", "vision", "ocrmypdf", "plugin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development :: Libraries",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]

[project.entry-points]


[project.urls]
Homepage = "https://github.com/mkyt/OCRmyPDF-AppleOCR"
Repository = "https://github.com/mkyt/OCRmyPDF-AppleOCR.git"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
ocrmypdf_appleocr = ["pdf.ttf"]


[dependency-groups]
dev = [
    "build>=1.3.0",
    "ruff>=0.14.0",
]

[tool.setuptools.dynamic]
version = { attr = "ocrmypdf_appleocr.__version__" }

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
